requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.4.0
xgboost>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
grad-cam>=1.5.0
pytesseract>=0.3.10
matplotlib>=3.7.0
git+https://github.com/openai/CLIP.git
